MAGE-1153 Perform additional refactoring
